feat: 添加全新 DB 组件 Chen, 去掉 OmniDB
