+++ src/app/elements/elements.component.ts
@@ -19,7 +19,7 @@ import {ElementReplayJsonComponent} from './replay/json/json.component';
 import {ElementConnectorKokoComponent} from './content/content-window/koko/koko.component';
 import {ElementConnectorMagnusComponent} from './content/content-window/magnus/magnus.component';
 import {ElementConnectorLionComponent} from './content/content-window/lion/lion.component';
-import {ElementConnectorOmnidbComponent} from './content/content-window/omnidb/omnidb.component';
+import {ElementConnectorChenComponent} from '@app/elements/content/content-window/chen/chen.component';
 import {ElementSelectAccountComponent} from './connect/connect-dialog/select-account/select-account.component';
 import {ElementReplayAsciicastComponent} from '@app/elements/replay/asciicast/asciicast.component';
 import {ElementAdvancedOptionComponent} from './connect/connect-dialog/advanced-option/advanced-option.component';
@@ -55,7 +55,7 @@ export const ElementComponents = [
   DisabledAssetsDialogComponent,
   ElementSettingComponent,
   ElementConnectDialogComponent,
-  ElementConnectorOmnidbComponent,
+  ElementConnectorChenComponent,
   ElementSelectAccountComponent,
   ElementReplayAsciicastComponent,
   ElementAdvancedOptionComponent,
